Fix PromptwareRoot resolution bug
